# Speed up add_image_checks