by pre-trimming down to the portion of the stmt that actually touches
input/output buffers.

On master, timing add_image_checks for a few apps:

local_laplacian: 16ms
lens_blur: 18ms
interpolate: 11ms
fft: 37ms

On this branch:

local_laplacian: 2.4ms
lens_blur: 11ms
interpolate: 3.5ms
fft: 3.9ms

This change was pulled from #3037