@abadams 2020-03-21
Only failing on the windows bots due to python issues. This change is not os-specific though, so I'm going to go ahead and merge.